Rewrite for brevity, emphasize applying function
The earlier version emphasizes only the ability of list comps to filter with `if` statements and doesn't actually change any of the values. I think applying some kind of change to each element is the more common use case for list comps and that's what I'd want to use this tag to teach.